requirements.txt
beautifulsoup4    == 4.8.2
googletrans       == 2.4.0
newspaper3k       == 0.2.8
nltk              == 3.4.5
numpy             == 1.18.0
pandas            == 0.25.3
requests          == 2.22.0
textblob          == 0.15.3
scikit-learn      == 0.20.1
notebook          == 6.0.2
matplotlib        == 3.1.2
seaborn           == 0.9.0
Flask             == 1.1.1